+++ Mollie/API/Resource/Chargebacks.py
@@ -1,0 +1,6 @@
+from .Base import *
+
+
+class Chargebacks(Base):
+    def getResourceObject(self, result):
+        return Chargeback(result)
