+++ Mollie/API/Resource/Refunds.py
@@ -1,0 +1,6 @@
+from .Base import *
+
+
+class Refunds(Base):
+    def getResourceObject(self, result):
+        return Refund(result)
