+++ Mollie/API/Resource/Subscriptions.py
@@ -10,7 +10,7 @@ class Subscriptions(Base):
     def getResourceObject(self, result):
         return Subscription(result)
 
-    def get(self, subscription_id):
+    def get(self, subscription_id, **params):
         if not subscription_id or not subscription_id.startswith(self.RESOURCE_ID_PREFIX):
             raise Error(
                 'Invalid subscription ID: "%s". A subscription ID should start with "%s".' % (subscription_id, self.RESOURCE_ID_PREFIX)
