+++ Mollie/API/Resource/__init__.py
@@ -1,5 +1,9 @@
 from .Base import *
 from .Payments import *
+from .Refunds import *
+from .Chargebacks import *
+from .PaymentRefunds import *
+from .PaymentChargebacks import *
 from .Issuers import *
 from .Methods import *
 from .Customers import *
